TFP-93: signed/unsigned fixup's

+++ src/bgp/inet6/test/inet6_table_test.cc
@@ -12,7 +12,7 @@
 #include "bgp/test/bgp_server_test_util.h"
 #include "control-node/control_node.h"
 
-static const int kRouteCount = 255;
+static const size_t kRouteCount = 255;
 
 class Inet6TableTest : public ::testing::Test {
 protected:
@@ -85,7 +85,7 @@ protected:
         Inet6Table::RequestKey key(prefix, NULL);
         Inet6Route *rt = dynamic_cast<Inet6Route *>(blue_->Find(&key));
         TASK_UTIL_EXPECT_TRUE(rt != NULL);
-        TASK_UTIL_EXPECT_EQ(1, rt->count());
+        TASK_UTIL_EXPECT_EQ(1U, rt->count());
         return true;
     }
 
@@ -136,35 +136,35 @@ TEST_F(Inet6TableTest, AddDeleteOneRoute) {
 TEST_F(Inet6TableTest, AddDeleteMultipleRoute1) {
     std::string ip_address = "2001:0db8:85a3::8a2e:0370:";
     std::string plen = "/128";
-    for (int idx = 1; idx <= kRouteCount; ++idx) {
+    for (size_t idx = 1; idx <= kRouteCount; ++idx) {
         std::ostringstream repr;
         repr << ip_address << idx << plen;
         AddRoute(repr.str());
     }
     task_util::WaitForIdle();
 
-    for (int idx = 1; idx <= kRouteCount; ++idx) {
+    for (size_t idx = 1; idx <= kRouteCount; ++idx) {
         std::ostringstream repr;
         repr << ip_address << idx << plen;
         VerifyRouteExists(repr.str());
     }
-    TASK_UTIL_EXPECT_EQ(adc_notification_, kRouteCount);
+    TASK_UTIL_EXPECT_EQ(static_cast<size_t>(adc_notification_), kRouteCount);
     TASK_UTIL_EXPECT_EQ(blue_->Size(), kRouteCount);
 
-    for (int idx = 1; idx <= kRouteCount; ++idx) {
+    for (size_t idx = 1; idx <= kRouteCount; ++idx) {
         std::ostringstream repr;
         repr << ip_address << idx << plen;
         DelRoute(repr.str());
     }
     task_util::WaitForIdle();
 
-    for (int idx = 1; idx <= kRouteCount; ++idx) {
+    for (size_t idx = 1; idx <= kRouteCount; ++idx) {
         std::ostringstream repr;
         repr << ip_address << idx << plen;
         VerifyRouteNoExists(repr.str());
     }
-    TASK_UTIL_EXPECT_EQ(del_notification_, kRouteCount);
-    TASK_UTIL_EXPECT_EQ(blue_->Size(), 0);
+    TASK_UTIL_EXPECT_EQ(static_cast<size_t>(del_notification_), kRouteCount);
+    TASK_UTIL_EXPECT_EQ(blue_->Size(), 0U);
 }
 
 // Add kRouteCount routes that differ only in the last byte i.e. the byte that
@@ -172,28 +172,28 @@ TEST_F(Inet6TableTest, AddDeleteMultipleRoute1) {
 TEST_F(Inet6TableTest, AddDeleteMultipleRoute2) {
     std::string ip_address = "2001:0db8:85a3:aaaa::0370:";
     std::string plen = "/64";
-    for (int idx = 1; idx <= kRouteCount; ++idx) {
+    for (size_t idx = 1; idx <= kRouteCount; ++idx) {
         std::ostringstream repr;
         repr << ip_address << idx << plen;
         AddRoute(repr.str());
     }
     task_util::WaitForIdle();
     TASK_UTIL_EXPECT_EQ(adc_notification_, 1);
-    TASK_UTIL_EXPECT_EQ(blue_->Size(), 1);
+    TASK_UTIL_EXPECT_EQ(blue_->Size(), 1U);
 
-    for (int idx = 1; idx <= kRouteCount; idx++) {
+    for (size_t idx = 1; idx <= kRouteCount; idx++) {
         std::ostringstream repr;
         repr << ip_address << (boost::format("%04X") % idx) << plen;
         DelRoute(repr.str());
     }
     TASK_UTIL_EXPECT_EQ(del_notification_, 1);
-    TASK_UTIL_EXPECT_EQ(blue_->Size(), 0);
+    TASK_UTIL_EXPECT_EQ(blue_->Size(), 0U);
 }
 
 TEST_F(Inet6TableTest, Hashing) {
     std::string ip_address = "2001:0db8:85a3:fedc:ba09:8a2e:0370:";
     std::string plen = "/128";
-    for (int idx = 1; idx <= kRouteCount; idx++) {
+    for (size_t idx = 1; idx <= kRouteCount; idx++) {
         std::ostringstream repr;
         repr << ip_address << (boost::format("%04X") % idx) << plen;
         AddRoute(repr.str());
@@ -203,15 +203,15 @@ TEST_F(Inet6TableTest, Hashing) {
     for (int idx = 0; idx < DB::PartitionCount(); idx++) {
         DBTablePartition *tbl_partition =
             static_cast<DBTablePartition *>(blue_->GetTablePartition(idx));
-        TASK_UTIL_EXPECT_NE(0, tbl_partition->size());
+        TASK_UTIL_EXPECT_NE(0U, tbl_partition->size());
     }
 
-    for (int idx = 1; idx <= kRouteCount; idx++) {
+    for (size_t idx = 1; idx <= kRouteCount; idx++) {
         std::ostringstream repr;
         repr << ip_address << (boost::format("%04X") % idx) << plen;
         DelRoute(repr.str());
     }
-    TASK_UTIL_EXPECT_EQ(blue_->Size(), 0);
+    TASK_UTIL_EXPECT_EQ(blue_->Size(), 0U);
 }
 
 int main(int argc, char **argv) {
